Add git-cl and git-try stubs.
They automatically git clone the original repository and execute the original script. So if they are first in the path, they will be executed. If a user has another git-cl/try in the path prior, these stubs won't be used at all and won't mess with the user's setup.

Also fix a bug with git handling in trychange.py
Review URL: http://codereview.chromium.org/119244

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@17737 4ff67af0-8c30-449e-8e8b-ad334ec8d88c