Make git-cl wrapper print errors instead of silently deleting repos.
It would previously delete and recreate git-cl-repo if it
saw that the remote was different than expected.  This looks
like it was added for the transition from the old SVN repo
to the new git one, but it also has the effect of losing
any local changes that have been made in the repo. :-(

Everyone should be updated by now, so print a helpful error
and exit instead.
Review URL: http://codereview.chromium.org/609002

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@39099 4ff67af0-8c30-449e-8e8b-ad334ec8d88c