Change scm.GIT.CaptureStatus to ignore unmerged upstream changes.
With this CL, the patch that is sent to the code review site includes only changes from the nearest common ancestor of the local and the upstream branch.

A diagram:

     /----B <- local branch
----A-----C <- upstream

Previously, the diff would be simply between C and B, whereas now it's between A and C, even if you update C.

This allows you to continously update origin/trunk without having to rebase/merge your local changes all the time, resulting in the need for less rebuilds.

Review URL: http://codereview.chromium.org/1583003

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@43113 4ff67af0-8c30-449e-8e8b-ad334ec8d88c