Add support for svn 1.4 to trychange.py
svn 1.4 doesn't have --depth parameter support so it cannot easily do a sparse checkout.

To workaround that, use svn import instead. It's also faster.
The main caveat is that the file must have a unique name, it uses a timestamp in file name for uniqueness.

Using a subdirectory (like the user name) doesn't work in the directory already exist.

It could be worked around but that would nullify the performance improvement.

Review URL: http://codereview.chromium.org/1783009

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@45811 4ff67af0-8c30-449e-8e8b-ad334ec8d88c