Fix git cl code to get the changelist description without using gcl.
The git_cl_hooks.py was failing to get the changelist description.
It was using gcl to get the description but this was failing
because gcl was unable to get the cached copy of 
codereview.settings (because it was never set up). In this
case, gcl reverts to the default dictionary which may be 
incorrect in some cases.

This changes to use 'git cl status --field=desc', which
is aware of the codereview.settings.

Review URL: http://codereview.chromium.org/2832006

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@49878 4ff67af0-8c30-449e-8e8b-ad334ec8d88c