Add chromite wrapper to depot_tools.
This file is a copy of the chromite wrapper that lives here:
  http://git.chromium.org/gitweb/?p=chromite.git;a=blob;f=bin/chromite

Anush requested that this wrapper be placed into depot_tools so
that users will get the wrapper in their path.  The wrapper will go
and find the real version of chromite based on the CWD.

Patch contributed by dianders@chromium.org.

Review URL: http://codereview.chromium.org/6299018

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@73011 4ff67af0-8c30-449e-8e8b-ad334ec8d88c