Partial update of upload.py with r660 patch.
Make json files known as a text mimetype so that they can be seen in reviews.

Original review at http://codereview.appspot.com/4133046/

Also make sure that the copy in third_party and in git_cl are the same. I'll
remove the copy in git_cl in a later change.

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@74116 4ff67af0-8c30-449e-8e8b-ad334ec8d88c