Enable back parallel sync on cygwin.
Cygwin used to have a bug where the number of lock was very limited, causing it
to crash when --jobs was used with a value different than 1. This bug seems to
have been fixed a few months ago as I can't reproduce the anymore after
extensive usage.

R=cmp@chromium.org
BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10689194

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@146798 4ff67af0-8c30-449e-8e8b-ad334ec8d88c