Add the ability to set the upstream branch
I can never remember the syntax (and always forget I have to specify the branch to set the upstream branch on) when using "git branch".

This patch just allows git-cl upstream to set the branch if an argument is provided. With no arg it prints the current one as it does today.

Review URL: https://chromiumcodereview.appspot.com/10825285

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@150945 4ff67af0-8c30-449e-8e8b-ad334ec8d88c