When a DEPS file is modified, run gclient sync.
It runs it by specifying BASE/HEAD to stay at the current revision.
This works around the lack of gclient sync when a DEPS file is applied on the
Try Server when triggered from Rietveld. This was handled my the Try Server
inside scripts/slave/chromium_command.py manually in the case a 'patch' is
specified.

TBR=rogerta@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/10918264

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@157110 4ff67af0-8c30-449e-8e8b-ad334ec8d88c