Fix toplevel_root parsing in trychange.py
In https://codereview.chromium.org/11574007/ I attempted to make it possible to use the TRYSERVER_ROOT setting in codereview.settings. It seems like some checkouts ran into errors with it not finding the codereview.settings file when it was parsed. This is probably due to the fact that the following call chain reads the self.toplevel_root variable:
_GclStyleSettings -> GetCodeReviewSetting -> ReadRootFile
By moving _GclStyleSettings up before self.toplevel_root is set probably causes this.

This CL attempts to correct this.

BUG=none, but try job upload fails for some users.
TEST=submitting try job to locally running try server.


Review URL: https://chromiumcodereview.appspot.com/11571052

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@173928 4ff67af0-8c30-449e-8e8b-ad334ec8d88c