Fix a crash in the owners check for a change without reviewers.
If we uploaded a CL w/o any reviewers specified, and then
ran the owners check on it (so that we had a Rietveld
issue number) we would crash in a Python assertion. This
wasn't caught in unit testing because of a limitation
in the unit test scaffolding.

TBR=maruel@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/12546003

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@186454 4ff67af0-8c30-449e-8e8b-ad334ec8d88c