Add reviewers to optional outputs for changes
This adds a comma-separated list of reviewers to the optional outputs
for changes. This can be added to the output using the custom formatter
option.  E.g. --output-format-changes="{url} {title} <{reviewers}>"

Also, for webkit changes, use CommitterList to map reviewer names to
emails so that they are similar format to other changes.


Review URL: https://chromiumcodereview.appspot.com/12446008

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@187351 4ff67af0-8c30-449e-8e8b-ad334ec8d88c