Adds --config option to download_from_google_storage.py
In order to support both developer workflow and not breaking bots, if the script
fails on a 403 in a bucket, it'll print a message asking developers to run
"download_from_google_storage --config" in order to create a new boto file.

This is not done automatically because it would break bots (Imagine hitting a 403,
and then gsutil wiping the .boto file, waiting for input, then dying).

BUG=231699,176331

Review URL: https://chromiumcodereview.appspot.com/17590010

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@208806 4ff67af0-8c30-449e-8e8b-ad334ec8d88c