Use plain sh instead of bash in the git template hooks.
The scripts are simple enough and should work fine under any POSIX shell,
which is particularly helpful when one is using `gclient sync' on a platform
that does not have /bin/bash (ie. it's installed in another location or not
present at all), since the hooks are automatically added to the git
checkouts.

While here, just `exec' the scripts directly instead of invoking another
shell by hand and remove the redundant `exit 0' lines.

R=cmp@chromium.org,dpranke@chromium.org,szager@chromium.org

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@214177 4ff67af0-8c30-449e-8e8b-ad334ec8d88c