Always convert flag value to string in set_flag
The callee needs it to be a string so that it can be made
part of HTTP request. However, in code using Rietveld
it's useful to use non-string types for flags where
Rietveld actually doesn't return strings (say booleans).

One example is the commit flag, which is a boolean.

BUG=291335

Review URL: https://chromiumcodereview.appspot.com/24846002

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@225748 4ff67af0-8c30-449e-8e8b-ad334ec8d88c