Fix "git cl format" when newer clang-format version is installed.
Newer versions of clang-format-diff.py now require a -i flag to
explicitely apply edits, otherwise they just print a diff, which
make "git cl format" a no-op.

This patch fixes the issue by probing the script's help text to
see if the flag is needed. If it is, it is added automatically.

BUG=NONE
R=maruel@chromium.org

Review URL: https://codereview.chromium.org/44263004

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@231020 4ff67af0-8c30-449e-8e8b-ad334ec8d88c