bypass_prodaccess fix
It looked like the bypass_prodaccess flag wasn't passed in all the way to
where it needed to be, causing it to be set to True by default at some point,
which caused the prodaccess path to never happen.  This change takes the flag
and brings it all of the way.

BUG=337972

Review URL: https://codereview.chromium.org/147443002

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@247049 4ff67af0-8c30-449e-8e8b-ad334ec8d88c