More fixes for hardcoded git directory paths.
The git path is based on the git version, so it might change whenever git is
updated. Run git.bat instead, since that's updated with new git installs and
should always point to the right path.

The mingw logic is based on similar handling of python.bat in git-cl.

Review URL: https://codereview.chromium.org/166273024

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@252584 4ff67af0-8c30-449e-8e8b-ad334ec8d88c