Add toolchain simple mirror option used for bots
I tried a git-ish tree/blob model, but way too slow for a ton of files.
This maps easily on to the current way the tree is mapped, saves
temp space on bots (because the full isos aren't downloaded), and is
a lot faster too. It means we can pull old revisions too now.

TBR=iannucci@chromium.org
BUG=348350

Review URL: https://codereview.chromium.org/185423004

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@254421 4ff67af0-8c30-449e-8e8b-ad334ec8d88c