Move Googlers to gs:// based toolchain pull
The gs:// download path is much faster. Move anyone who could have
access to gs://chrome-wintoolchain/ over to that for less pain.

The downside is that we have to prompt them to run
download_from_google_storage --config which people are going to
complain about, but I think this is a better experience overall.

R=iannucci@chromium.org,hinoka@chromium.org
BUG=349964

Review URL: https://codereview.chromium.org/191373002

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@255696 4ff67af0-8c30-449e-8e8b-ad334ec8d88c