Accept remote urls as gospel in git_cache.
git_cache rewrites all urls to https://netloc/path when
setting them as remote urls. This doesn't work when the
remote is actually local on disk, and the "url" is a
/path/to/repo. This seems like an unlikely case, but it
is leveraged by testing, and it doesn't hurt to support
it, so now we do.

R=iannucci@chromium.org
BUG=261619

Review URL: https://codereview.chromium.org/196493002

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@256386 4ff67af0-8c30-449e-8e8b-ad334ec8d88c