Fix the gn wrapper to be aware of the --root argument.
The gn binary itself is capable of being run from outside of a
checkout as long as you pass the --root flag.

However, the gn.py wrapper script needs to *also* know where the the
checkout is, in order to figure out how to find the gn binary itself.

This patch changes the wrapper to be aware of the --root arg :).

R=brettw@chromium.org
BUG=

Review URL: https://codereview.chromium.org/212173004

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@259998 4ff67af0-8c30-449e-8e8b-ad334ec8d88c