Use get_or_create_merge_base in git-cl-upload to correctly deal with rebased upstreams.
In the event that users are NOT using the enhanced tooling, this is equivalent
to `git merge-base "@{u}" HEAD`. In the event that they are, this will catch
the case where their parent branch got rebased, but this branch hasn't been
rebased on the parent yet.

R=agable@chromium.org, szager@chromium.org
BUG=

Review URL: https://codereview.chromium.org/224703002

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@261601 4ff67af0-8c30-449e-8e8b-ad334ec8d88c