Put patches to refs/patches/<user>/<name>
When using Git, trychange.py used to put patches to refs/patches/<name>
refs, where <name> is a name of a patch. Now the using the username as a part
of the path.

This way we have better control over permissions per user basis

R=agable@chromium.org, iannucci@chromium.org
TEST=Sent a change to quickoffice try server. Master and slave do not
     need to be modified because the ref file in master branch points to
     the correct branch name

Review URL: https://codereview.chromium.org/231663005

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@263432 4ff67af0-8c30-449e-8e8b-ad334ec8d88c