Do not fetch the CL description when using git cl status --fast.
This causes a network I/O, which is not fast. Also this is a non-starter in the
case of no network connectivity, causing a crash instead of completing
successfully. Until Air Canada provides on-flight internet, this is a
non-starter. #firstworldproblem.

Also fix git cl presubmit in case of lack of network connectivity.

R=stip@chromium.org
BUG=

Review URL: https://codereview.chromium.org/418293004

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@285911 4ff67af0-8c30-449e-8e8b-ad334ec8d88c