Add git cherry picking extension
This extension uploads a fake cherry pick-style diff to rietveld with a modified project parameter. The modified project is intended to be used by the commit queue to attempt to land the change on a branch.

This works by grabbing the parent of the targeted revision and generating the diff. It is intended to be used to CQ trivial cherry picks which apply cleanly on top of other branches without conflicts.

BUG=387111

Review URL: https://codereview.chromium.org/397593004

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@286273 4ff67af0-8c30-449e-8e8b-ad334ec8d88c