Use a regex instead of shlex.split() to get remote url
shlex.spit(), in addition to doing str.split(' '), also strips out '\\'
from windows paths, which causes the cache_dir check to fail and Gclient to think
that we're not in cache_dir mode even if we are.

Instead of using shlex to strip data from a stdout log, I think regex is more
suited for this job.

BUG=405973

Review URL: https://codereview.chromium.org/497053002

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@291446 4ff67af0-8c30-449e-8e8b-ad334ec8d88c